Agents!
I got a very rudimentary, hard-coded, version of the agents working. Now, agents will buy homes and they will appear on the map as expected. This is very rough, and the agents will buy them on any LandSpace type, not just Residential spaces - but this is the beginning of it.