Selection cleanup
I started to clean up a few things as far as input mode and selection are concerned. It's not finished, but I'm heading in the right direction. I also added a rudimentary Information Popup that will show you information about a single "selected" land. Need to work on that a bit more yet.