v0.065 - UI Cleanup
So after some more noodling, I think I've gotten to a state with the UI that is "good enough" for now. All of the things I want working are there, and for the most part it feels good. However, looking at my backlog everything looks kinda of squishy and not well defined. So, I'm going to take some time to go through it and decide what to do next.
I'm probably not going to make any fancy graphics and instead prototype the next set of features so that I can ultimately decide if this idea is going to be fun or not. I hope it is.