Decision Engine
I integrated my scratch Behavior Tree work with the game. Now, the game will check conditions and follow the simple Decision Tree I've got set up. The end result is that popups ask the player question, and depending on their choice, it will affect the game in some way.

It's, as usual, a bit rough around the edges, but I think it's a decent base to work from going forward. Now, I just need to flesh out a much more detailed actual Decision Tree for the real game. I'll get to that someday...