Game features!
I had been thinking about how exactly I was going to accomplish this for awhile, but I think I've nailed down a scalable way to do it. So...what this is, is a way for individual "features" in the game to be turned on/off/adjusted as actions in the decision tree happen. So, in this POC, none of the residential/commercial/industrial/agricultural zones are turned on, until you start to make decisions as they pop up.

This is very rough, and needs refinement, but I think it'll work for the game going forward. Committing now before I think better of it and change something.