Behavior Tree bug
There was a bug in my Behavior Tree implementation that would allow branches to execute if their criteria matched the current input. So...for any Behavior Node that had an Entrance Criteria of "No," whenever a Dialog answer was "No" all branches that had that as an Entrance Criteria would execute. Was a real head scratcher as I had looked at the code for far too long.

Either way, now the leaf nodes that aren't executed are removed from the Behavior Tree so that this doesn't happen. I implemented another node in the Tree and added another GameFeature to test this out. Seems to be working. 

Hopefully next I'll add a few more nodes to the Behavior Tree in the Decision Engine, and then move on to adding some sort of alerts to the game.