Menu animation progress
I got the "sliding in" animation started here, but things are acting a little odd. It's almost as if the scene instances aren't independent and are sharing child nodes for some reason. Either way, the vertical slide in works, but the right slide in is using the vertical animation for some reason.