Menu buildout.
I started to build out the menu as it might look in the MVP of the game, and am adding in features as needed. Right now the enabled/disabled status of the buttons are starting to be wired up, as well as any actions the buttons might need to be perform. Still a ways to go though.