Polishing menu/features.
I've continued to chip away at features for the menu to the point where gluing new things on isn't so bad. It more or less works like I had envisioned, although there is a bit more polish to add when the time comes.

I also started back on my todo list for getting to an MVP. Next up will be placing the players official house somewhere on the map.